Step 4 -> Inheritance.
We can create classes that are extensions or specializations of other classes.
Objects or instances of such specialized classes will be considered instances of the former classes, but the opposite will not be true.

We created a new MotifTrack class that inherits the functionalities of the Track class but now the notes are not fired randomly from the scale notes. Instead, we have a motif that dictates the notes that are fired ;)